Basic fretboard models and views
Capability to create a fretboard with arbitrary number of frets, and
tuning by open note name.
Strings are modeled as collections of frets with a root note, and the
fretboard is modeled as a collection of strings.  Probably a bad
definition for a fret, but not sure what else to call it.